Module needletail::errors

The errors needletail can return; only when parsing FASTA/FASTQ files

Structs

ErrorPosition

Represents where we were in a file when an error occurred.

ParseError

The only error type that needletail returns

Enums

ParseErrorKind

The type of error that occured during file parsing