Struct needletail::parser::SequenceRecord

pub struct SequenceRecord<'a> { /* private fields */ }

A FASTA or FASTQ record

Implementations

impl<'a> SequenceRecord<'a>

pub fn format(&self) -> Format

Returns the format of the record

pub fn id(&self) -> &[u8]

Returns the id of the record

pub fn raw_seq(&self) -> &[u8]

Returns the raw sequence of the record. Only matters for FASTA since it can contain newlines.

pub fn seq(&self) -> Cow<'_, [u8]>

Returns the cleaned up sequence of the record. For FASTQ it is the same as raw_seq but for FASTA it is raw_seq minus all the \r\n

pub fn qual(&self) -> Option<&[u8]>

Returns the quality line if there is one. Always None for FASTA and Some for FASTQ, even if the quality line is empty.

pub fn all(&self) -> &[u8]

Returns the full sequence, including line endings. This doesn’t include a trailing newline.

pub fn num_bases(&self) -> usize

Return the number of bases in the sequence, computed efficiently.

pub fn start_line_number(&self) -> u64

Return the line number in the file of the start of the sequence

pub fn line_ending(&self) -> LineEnding

Which line ending is this record using?

pub fn write( &self, writer: &mut dyn Write, forced_line_ending: Option<LineEnding>, ) -> Result<(), ParseError>

Write record back to a Write instance. By default it will use the original line ending but you can force it to use another one.