Module needletail::sequence

Generic functions for working with (primarily nucleic acid) sequences

Traits

QualitySequence

[⚠️Unstable] A trait to wrap over sequence data that has associated quality information.

Sequence

A generic FASTX record that also abstracts over several logical operations that can be performed on nucleic acid sequences.

Functions

canonical

Taking in a sequence string, return the canonical form of the sequence (e.g. the lexigraphically lowest of either the original sequence or its reverse complement)

complement

Returns the complementary base for a given IUPAC base code.

minimizer

Find the lexigraphically smallest substring of seq of length length

normalize

Transform a nucleic acid sequence into its “normalized” form.